Refactor common DRF processing into Pydantic model
