+++ gateway/sds_gateway/api_methods/tasks.py
@@ -613,7 +613,9 @@ def notify_shared_users(
 
 
 @shared_task
-def start_capture_post_processing(capture_uuid: str, processing_config: dict) -> dict:
+def start_capture_post_processing(
+    capture_uuid: str, processing_config: dict[str, dict[str, Any]]
+) -> dict[str, Any]:
     """
     Start post-processing pipeline for a DigitalRF capture.
 
@@ -624,8 +626,8 @@ def start_capture_post_processing(capture_uuid: str, processing_config: dict) ->
         capture_uuid: UUID of the capture to process
         processing_config: Dict with processing configurations, e.g.:
             {
-                "spectrogram": {"fft_size": 1024, "std_dev": 100, "hop_size": 500, "colormap": "magma"},
-                "waterfall": {...}
+                "spectrogram": {"fft_size": 1024, "std_dev": 100, "hop_size": 500,
+                "colormap": "magma"}, "waterfall": {...}
             }
     """
     logger.info(f"Starting post-processing pipeline for capture {capture_uuid}")
@@ -652,30 +654,21 @@ def start_capture_post_processing(capture_uuid: str, processing_config: dict) ->
         # Launch the visualization pipeline with processing config
         # Individual cogs will check if they should run based on processing_config
         logger.info(
-            f"Launching pipeline {pipeline_name} for capture {capture_uuid} with config: {processing_config}"
+            f"Launching pipeline {pipeline_name} for capture {capture_uuid} "
+            f"with config: {processing_config}"
         )
-        try:
-            pipeline.launch(
-                capture_uuid=capture_uuid, processing_config=processing_config
-            )
-            logger.info(f"Pipeline launched successfully for capture {capture_uuid}")
-        except Exception as e:
-            logger.error(f"Pipeline launch failed for capture {capture_uuid}: {e}")
-            raise
+        pipeline.launch(capture_uuid=capture_uuid, processing_config=processing_config)
+        logger.info(f"Pipeline launched successfully for capture {capture_uuid}")
 
         return {
             "status": "success",
             "message": (
-                f"Post-processing pipeline started for {len(processing_config)} processing types"
+                f"Post-processing pipeline started for {len(processing_config)} "
+                f"processing types"
             ),
             "capture_uuid": capture_uuid,
             "processing_config": processing_config,
         }
-
-    except Capture.DoesNotExist:
-        error_msg = f"Capture {capture_uuid} not found"
-        logger.error(error_msg)
-        raise
     except Exception as e:
         error_msg = f"Unexpected error in post-processing pipeline: {e}"
         logger.exception(error_msg)
