+++ gateway/sds_gateway/visualizations/processing/utils.py
@@ -5,13 +5,144 @@ import os
 from pathlib import Path
 from typing import Any
 
+import h5py
+from digital_rf import DigitalRFReader
 from django.conf import settings
 from loguru import logger
+from pydantic import BaseModel
+from pydantic import Field
+from pydantic import computed_field
+from pydantic import field_validator
+from pydantic import model_validator
 
-from sds_gateway.api_methods.models import Capture
 from sds_gateway.api_methods.utils.minio_client import get_minio_client
-from sds_gateway.visualizations.models import PostProcessedData
-from sds_gateway.visualizations.models import ProcessingStatus
+
+
+class DigitalRFParams(BaseModel):
+    """Base parameters for processing DigitalRF data with validation."""
+
+    reader: Any = Field(exclude=True)  # Exclude from serialization
+    channel: str = Field(min_length=1, description="Channel name")
+    start_sample: int = Field(ge=0, description="Start sample index")
+    end_sample: int = Field(gt=0, description="End sample index")
+    fft_size: int = Field(gt=0, description="FFT size")
+    center_freq: float = Field(description="Center frequency in Hz")
+    sample_rate_numerator: int = Field(gt=0, description="Sample rate numerator")
+    sample_rate_denominator: int = Field(gt=0, description="Sample rate denominator")
+
+    @field_validator("end_sample")
+    @classmethod
+    def validate_end_sample(cls, v, info):
+        """Validate that end_sample is greater than start_sample."""
+        if "start_sample" in info.data and v <= info.data["start_sample"]:
+            msg = "end_sample must be greater than start_sample"
+            raise ValueError(msg)
+        return v
+
+    @model_validator(mode="after")
+    def validate_channel_exists(self):
+        """Validate that the channel exists in the DigitalRF data."""
+        channels = self.reader.get_channels()
+        if not channels:
+            msg = "No channels found in DigitalRF data"
+            raise ValueError(msg)
+        if self.channel not in channels:
+            msg = (
+                f"Channel {self.channel} not found in DigitalRF data. "
+                f"Available channels: {channels}"
+            )
+            raise ValueError(msg)
+        return self
+
+    @computed_field
+    @property
+    def sample_rate(self) -> float:
+        """Calculate sample rate from numerator and denominator."""
+        return float(self.sample_rate_numerator) / float(self.sample_rate_denominator)
+
+    @computed_field
+    @property
+    def min_frequency(self) -> float:
+        """Calculate minimum frequency."""
+        return self.center_freq - self.sample_rate / 2
+
+    @computed_field
+    @property
+    def max_frequency(self) -> float:
+        """Calculate maximum frequency."""
+        return self.center_freq + self.sample_rate / 2
+
+    @computed_field
+    @property
+    def total_samples(self) -> int:
+        """Calculate total number of samples."""
+        return self.end_sample - self.start_sample
+
+    class Config:
+        arbitrary_types_allowed = True  # Allow DigitalRFReader type
+
+
+def validate_digitalrf_data(
+    drf_path: Path, channel: str, fft_size: int = 1024
+) -> DigitalRFParams:
+    """
+    Load DigitalRF data and return DigitalRFParams with validated attributes.
+
+    This function loads DigitalRF data and creates a validated DigitalRFParams
+    instance. Pydantic handles all validation logic declaratively.
+
+    Args:
+        drf_path: Path to DigitalRF data directory
+        channel: Channel name to process
+
+    Returns:
+        DigitalRFParams: Validated parameters for DigitalRF processing
+
+    Raises:
+        ValidationError: If DigitalRF data is invalid or missing required information
+    """
+    # Initialize DigitalRF reader
+    reader = DigitalRFReader(str(drf_path))
+
+    # Get sample bounds - let this fail naturally if invalid
+    bounds = reader.get_bounds(channel)
+    if bounds is None:
+        msg = "Could not get sample bounds for channel"
+        raise ValueError(msg)
+    start_sample, end_sample = bounds
+    if start_sample is None or end_sample is None:
+        msg = "Invalid sample bounds for channel"
+        raise ValueError(msg)
+
+    # Get metadata from DigitalRF properties
+    drf_props_path = drf_path / channel / "drf_properties.h5"
+    with h5py.File(drf_props_path, "r") as f:
+        sample_rate_numerator = f.attrs.get("sample_rate_numerator")
+        sample_rate_denominator = f.attrs.get("sample_rate_denominator")
+        if sample_rate_numerator is None or sample_rate_denominator is None:
+            msg = "Sample rate information missing from DigitalRF properties"
+            raise ValueError(msg)
+
+    # Get center frequency from metadata (optional)
+    center_freq = 0.0
+    try:
+        metadata_dict = reader.read_metadata(start_sample, end_sample, channel)
+        if metadata_dict and "center_freq" in metadata_dict:
+            center_freq = float(metadata_dict["center_freq"])
+    except Exception as e:  # noqa: BLE001
+        logger.warning(f"Could not read center frequency from metadata: {e}")
+
+    # Create DigitalRFParams
+    return DigitalRFParams(
+        reader=reader,
+        channel=channel,
+        start_sample=start_sample,
+        end_sample=end_sample,
+        fft_size=fft_size,
+        center_freq=center_freq,
+        sample_rate_numerator=sample_rate_numerator,
+        sample_rate_denominator=sample_rate_denominator,
+    )
 
 
 def reconstruct_drf_files(capture, capture_files, temp_path: Path) -> Path | None:
@@ -77,6 +208,8 @@ def store_processed_data(
     Returns:
         dict: Storage result
     """
+    from sds_gateway.api_methods.models import Capture  # noqa: PLC0415
+    from sds_gateway.visualizations.models import PostProcessedData  # noqa: PLC0415
 
     logger.info(f"Storing {processing_type} file for capture {capture_uuid}")
 
@@ -119,7 +252,7 @@ def store_processed_data(
 
 
 def create_or_reset_processed_data(
-    capture, processing_type: str, processing_parameters: dict | None = None
+    capture, processing_type: str, processing_parameters: dict[str, Any] | None = None
 ):
     """Create or reset a PostProcessedData record for a capture and processing type.
 
@@ -132,11 +265,13 @@ def create_or_reset_processed_data(
         PostProcessedData: The created or reset record
     """
     # Import models here to avoid Django app registry issues
-    from sds_gateway.visualizations.models import PostProcessedData
+    from sds_gateway.visualizations.models import PostProcessedData  # noqa: PLC0415
+    from sds_gateway.visualizations.models import ProcessingStatus  # noqa: PLC0415
 
     # Try to get existing record (any record for this capture and processing type)
     logger.info(
-        f"Looking for existing PostProcessedData for capture {capture.uuid}, type {processing_type}"
+        f"Looking for existing PostProcessedData for capture {capture.uuid}, "
+        f"type {processing_type}"
     )
     processed_data = PostProcessedData.objects.filter(
         capture=capture,
